updated validation and role guards for customer module

+++ src/customers/customers.controller.ts
@@ -9,9 +9,9 @@ import {
   Delete,
   UseGuards,
   Param,
-  HttpException,
-  HttpStatus,
   Query,
+  NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { CustomersService } from './customers.service';
 import { CreateCustomerDto } from './dto/create-customer.dto';
@@ -21,6 +21,7 @@ import { Roles } from 'src/auth/decorators/roles.decorator';
 import { Role } from '@prisma/client';
 import { RoleGuard } from 'src/auth/role.guard';
 import { CurrentUser } from 'src/auth/decorators/user.decorator';
+import { JWTPayload } from 'src/auth/types/auth.type';
 
 @Controller('customers')
 export class CustomersController {
@@ -36,18 +37,20 @@ export class CustomersController {
   @Roles(Role.Admin)
   @UseGuards(RoleGuard)
   @Get()
-  findAll(@Query() query: { firstName: string; lastName: string }) {
+  findAll(@Query() query: { firstName?: string; lastName?: string }) {
     return this.customersService.findAll(query.firstName, query.lastName);
   }
 
-  @Roles(Role.Admin, Role.Customer, Role.Staff)
-  @UseGuards(RoleGuard)
   @Get('/:id')
-  async findOne(@Param('id') id: string) {
+  async findOne(@CurrentUser() user: JWTPayload, @Param('id') id: string) {
+    if (user.role === Role.Customer && id !== user.id) {
+      throw new ForbiddenException();
+    }
+
     const customer = await this.customersService.findOneById(id);
 
     if (!customer) {
-      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+      throw new NotFoundException('Customer not found');
     }
     return customer;
   }
@@ -56,7 +59,7 @@ export class CustomersController {
   @UseGuards(RoleGuard)
   @Patch('/current')
   update(
-    @CurrentUser('uid') id: string,
+    @CurrentUser('id') id: string,
     @Body() updateCustomerDto: UpdateCustomerDto,
   ) {
     return this.customersService.update(id, updateCustomerDto);
@@ -64,7 +67,11 @@ export class CustomersController {
 
   @Roles(Role.Customer, Role.Admin)
   @Delete('/:id')
-  remove(@Param('id') id: string) {
+  remove(@CurrentUser() user: JWTPayload, @Param('id') id: string) {
+    if (user.role === Role.Customer && id !== user.id) {
+      throw new ForbiddenException();
+    }
+
     return this.customersService.remove(id);
   }
 }
